IDC: move unit tests from Jetpack to the IDC package.
