+++ projects/packages/identity-crisis/changelog/update-move_idc_tests
@@ -1,0 +1,5 @@
+Significance: patch
+Type: added
+Comment: Move unit tests from Jetpack to the identity-crisis package.
+
+
