+++ projects/plugins/jetpack/tests/php/general/test_class.jetpack.php
@@ -3,7 +3,6 @@
 use Automattic\Jetpack\Connection\Manager as Connection_Manager;
 use Automattic\Jetpack\Constants;
 use Automattic\Jetpack\Assets;
-use Automattic\Jetpack\Identity_Crisis;
 use Automattic\Jetpack\Partner;
 use Automattic\Jetpack\Status;
 
@@ -371,129 +370,6 @@ EXPECTED;
 		$this->assertFalse( Jetpack::get_other_linked_admins() );
 	}
 
-	function test_idc_optin_default() {
-		if ( is_multisite() ) {
-			$this->assertFalse( Identity_Crisis::sync_idc_optin() );
-		} else {
-			$this->assertTrue( Identity_Crisis::sync_idc_optin() );
-		}
-	}
-
-	function test_idc_optin_filter_overrides_development_version() {
-		add_filter( 'jetpack_development_version', '__return_true' );
-		add_filter( 'jetpack_sync_idc_optin', '__return_false' );
-		$this->assertFalse( Identity_Crisis::sync_idc_optin() );
-		remove_filter( 'jetpack_development_version', '__return_true' );
-		remove_filter( 'jetpack_sync_idc_optin', '__return_false' );
-	}
-
-	function test_idc_optin_casts_to_bool() {
-		add_filter( 'jetpack_sync_idc_optin', array( $this, 'return_string_1' ) );
-		$this->assertTrue( Identity_Crisis::sync_idc_optin() );
-		remove_filter( 'jetpack_sync_idc_optin', array( $this, 'return_string_1' ) );
-	}
-
-	function test_idc_optin_true_when_constant_true() {
-		Constants::set_constant( 'JETPACK_SYNC_IDC_OPTIN', true );
-		$this->assertTrue( Identity_Crisis::sync_idc_optin() );
-	}
-
-	function test_idc_optin_false_when_constant_false() {
-		Constants::set_constant( 'JETPACK_SYNC_IDC_OPTIN', false );
-		$this->assertFalse( Identity_Crisis::sync_idc_optin() );
-	}
-
-	function test_idc_optin_filter_overrides_constant() {
-		Constants::set_constant( 'JETPACK_SYNC_IDC_OPTIN', true );
-		add_filter( 'jetpack_sync_idc_optin', '__return_false' );
-		$this->assertFalse( Identity_Crisis::sync_idc_optin() );
-		remove_filter( 'jetpack_sync_idc_optin', '__return_false' );
-	}
-
-	function test_sync_error_idc_validation_returns_false_if_no_option() {
-		Jetpack_Options::delete_option( 'sync_error_idc' );
-		$this->assertFalse( Identity_Crisis::validate_sync_error_idc_option() );
-	}
-
-	function test_sync_error_idc_validation_returns_true_when_option_matches_expected() {
-		add_filter( 'jetpack_sync_idc_optin', '__return_true' );
-		Jetpack_Options::update_option( 'sync_error_idc', Identity_Crisis::get_sync_error_idc_option() );
-		$this->assertTrue( Identity_Crisis::validate_sync_error_idc_option() );
-		Jetpack_Options::delete_option( 'sync_error_idc' );
-		remove_filter( 'jetpack_sync_idc_optin', '__return_true' );
-	}
-
-	/**
-	 * Verify that validate_sync_error returns false if wpcom_ is set and matches expected.
-	 */
-	public function test_sync_error_idc_validation_returns_false_when_wpcom_option_matches_expected() {
-		add_filter( 'jetpack_sync_idc_optin', '__return_true' );
-		$option                  = Identity_Crisis::get_sync_error_idc_option();
-		$option['wpcom_home']    = $option['home'];
-		$option['wpcom_siteurl'] = $option['siteurl'];
-		Jetpack_Options::update_option( 'sync_error_idc', $option );
-		$this->assertFalse( Identity_Crisis::validate_sync_error_idc_option() );
-
-		// Verify the migrate_for_idc is set.
-		$this->assertTrue( Jetpack_Options::get_option( 'migrate_for_idc' ) );
-
-		Jetpack_Options::delete_option( 'sync_error_idc' );
-		Jetpack_Options::delete_option( 'migrate_for_idc' );
-		remove_filter( 'jetpack_sync_idc_optin', '__return_true' );
-	}
-
-	/**
-	 * Verify that validate_sync_error returns true if wpcom_ is set and does not match.
-	 */
-	public function test_sync_error_idc_validation_returns_true_when_wpcom_option_does_not_match_expected() {
-		add_filter( 'jetpack_sync_idc_optin', '__return_true' );
-		$option                  = Identity_Crisis::get_sync_error_idc_option();
-		$option['wpcom_home']    = $option['home'];
-		$option['wpcom_siteurl'] = 'coolrunnings.test';
-		Jetpack_Options::update_option( 'sync_error_idc', $option );
-		$this->assertTrue( Identity_Crisis::validate_sync_error_idc_option() );
-
-		// Verify the migrate_for_idc is not set.
-		$this->assertNotTrue( Jetpack_Options::get_option( 'migrate_for_idc' ) );
-
-		Jetpack_Options::delete_option( 'sync_error_idc' );
-		Jetpack_Options::delete_option( 'migrate_for_idc' );
-		remove_filter( 'jetpack_sync_idc_optin', '__return_true' );
-	}
-
-	function test_sync_error_idc_validation_cleans_up_when_validation_fails() {
-		Jetpack_Options::update_option( 'sync_error_idc', array(
-			'home'    => 'coolsite.com/',
-			'siteurl' => 'coolsite.com/wp/',
-		) );
-
-		$this->assertFalse( Identity_Crisis::validate_sync_error_idc_option() );
-		$this->assertFalse( Jetpack_Options::get_option( 'sync_error_idc' ) );
-	}
-
-	function test_sync_error_idc_validation_cleans_up_when_part_of_validation_fails() {
-		$test            = Identity_Crisis::get_sync_error_idc_option();
-		$test['siteurl'] = 'coolsite.com/wp/';
-		Jetpack_Options::update_option( 'sync_error_idc', $test );
-
-		$this->assertFalse( Identity_Crisis::validate_sync_error_idc_option() );
-		$this->assertFalse( Jetpack_Options::get_option( 'sync_error_idc' ) );
-	}
-
-	function test_sync_error_idc_validation_returns_false_and_cleans_up_when_opted_out() {
-		Jetpack_Options::update_option( 'sync_error_idc', Identity_Crisis::get_sync_error_idc_option() );
-		Constants::set_constant( 'JETPACK_SYNC_IDC_OPTIN', false );
-
-		$this->assertFalse( Identity_Crisis::validate_sync_error_idc_option() );
-		$this->assertFalse( Jetpack_Options::get_option( 'sync_error_idc' ) );
-	}
-
-	function test_is_staging_site_true_when_sync_error_idc_is_valid() {
-		add_filter( 'jetpack_sync_error_idc_validation', '__return_true' );
-		$this->assertTrue( ( new Status() )->is_staging_site() );
-		remove_filter( 'jetpack_sync_error_idc_validation', '__return_false' );
-	}
-
 	function test_is_dev_version_true_with_alpha() {
 		Constants::set_constant( 'JETPACK__VERSION', '4.3.1-alpha' );
 		$this->assertTrue( Jetpack::is_development_version() );
@@ -539,44 +415,6 @@ EXPECTED;
 		add_filter( 'jetpack_offline_mode', '__return_zero' );
 		$this->assertFalse( ( new Status() )->is_offline_mode() );
 		remove_filter( 'jetpack_offline_mode', '__return_zero' );
-	}
-
-	function test_get_sync_idc_option_sanitizes_out_www_and_protocol() {
-		$original_home    = get_option( 'home' );
-		$original_siteurl = get_option( 'siteurl' );
-
-		update_option( 'home', 'http://www.coolsite.com' );
-		update_option( 'siteurl', 'http://www.coolsite.com/wp' );
-
-		$expected = array(
-			'home' => 'coolsite.com/',
-			'siteurl' => 'coolsite.com/wp/'
-		);
-
-		$this->assertSame( $expected, Identity_Crisis::get_sync_error_idc_option() );
-
-		// Cleanup
-		update_option( 'home', $original_home );
-		update_option( 'siteurl', $original_siteurl );
-	}
-
-	function test_get_sync_idc_option_with_ip_address_in_option() {
-		$original_home    = get_option( 'home' );
-		$original_siteurl = get_option( 'siteurl' );
-
-		update_option( 'home', 'http://72.182.131.109/~wordpress' );
-		update_option( 'siteurl', 'http://72.182.131.109/~wordpress/wp' );
-
-		$expected = array(
-			'home' => '72.182.131.109/~wordpress/',
-			'siteurl' => '72.182.131.109/~wordpress/wp/'
-		);
-
-		$this->assertSame( $expected, Identity_Crisis::get_sync_error_idc_option() );
-
-		// Cleanup
-		update_option( 'home', $original_home );
-		update_option( 'siteurl', $original_siteurl );
 	}
 
 	function test_normalize_url_protocol_agnostic_strips_protocol_and_www_for_subdir_subdomain() {
@@ -797,15 +635,6 @@ EXPECTED;
 			$password .= $password;
 		}
 		return $password;
-	}
-
-	/**
-	 * Return string '1'.
-	 *
-	 * @return string
-	 */
-	public function return_string_1() {
-		return '1';
 	}
 
 	/**
